requirements.txt
# PrivateEdge — Python stack
onnxruntime>=1.20.0
django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
channels>=4.0
daphne>=4.0
numpy>=1.26
opencv-python-headless>=4.9
mss>=9.0
sounddevice>=0.5
faster-whisper>=1.0.0
pyvirtualcam>=0.11.0
pytest>=8.0

# Optional NSFW via Hugging Face (set PRIVATEEDGE_USE_HF_EFFICIENTNET=1)
torch>=2.0.0
transformers>=4.36.0
Pillow>=10.0.0
